fix golint check
